# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: WIAL Backend API
  version: 1.0.0
  description: |
    OpenAPI documentation for the WIAL backend.

    Public auth is login-only. User creation is restricted to authenticated
    administrators through the managed users endpoint.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/team-members:
    get:
      tags: [Team Members]
      summary: List chapter team members
      operationId: listTeamMembers
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Team member list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Team Members]
      summary: Create a chapter team member
      operationId: createTeamMember
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberCreateRequest'
      responses:
        '201':
          description: Team member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/team-members/{id}:
    get:
      tags: [Team Members]
      summary: Get a chapter team member
      operationId: getTeamMember
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Team member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Team Members]
      summary: Patch a chapter team member
      operationId: patchTeamMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberPatchRequest'
      responses:
        '200':
          description: Team member patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Team Members]
      summary: Delete a chapter team member
      operationId: deleteTeamMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Team member deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/resources:
    get:
      tags: [Resources]
      summary: List resources
      description: Omitting `chapter_id` returns global resources only.
      operationId: listResources
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Resources]
      summary: Create a resource
      description: Super-admins can create global or chapter resources. Chapter leads can create resources only for their chapter.
      operationId: createResource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/resources/{id}:
    get:
      tags: [Resources]
      summary: Get a resource
      operationId: getResource
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Resource found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Resources]
      summary: Patch a resource
      operationId: patchResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourcePatchRequest'
      responses:
        '200':
          description: Resource patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Resources]
      summary: Delete a resource
      operationId: deleteResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Resource deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role or assignment scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TeamMember:
      type: object
      properties:
        id: { type: string }
        chapter_id: { type: string }
        name: { type: string }
        role: { type: string }
        blurb: { type: string }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, chapter_id, name, role, blurb, sort_order, created_at, updated_at]
    TeamMemberCreateRequest:
      type: object
      properties:
        chapter_id: { type: string }
        name: { type: string }
        role: { type: string }
        blurb: { type: string }
        sort_order: { type: integer, default: 0 }
      required: [chapter_id, name, role, blurb]
    TeamMemberPatchRequest:
      type: object
      properties:
        chapter_id: { type: string }
        name: { type: string }
        role: { type: string }
        blurb: { type: string }
        sort_order: { type: integer }
    TeamMemberListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
      required: [data]
    Resource:
      type: object
      properties:
        id: { type: string }
        chapter_id: { type: string, nullable: true }
        title: { type: string }
        type: { type: string }
        summary: { type: string }
        url: { type: string, nullable: true }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, title, type, summary, sort_order, created_at, updated_at]
    ResourceCreateRequest:
      type: object
      properties:
        chapter_id: { type: string, nullable: true }
        title: { type: string }
        type: { type: string }
        summary: { type: string }
        url: { type: string, nullable: true }
        sort_order: { type: integer, default: 0 }
      required: [title, type, summary]
    ResourcePatchRequest:
      type: object
      properties:
        chapter_id: { type: string }
        title: { type: string }
        type: { type: string }
        summary: { type: string }
        url: { type: string }
        sort_order: { type: integer }
    ResourceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
      required: [data]
...